Remove unused AMC driver files * Removed amc duplicate files in ami driver * Readme update

+++ sw/AMI/driver/ami_pcie.c
@@ -1026,8 +1026,8 @@ static int do_bar_transaction(struct pci_dev	*dev,
 	bar = &(pf_dev->pcie_config->header->bar[bar_idx]);
 
 	if ((bar->len == 0) || ((offset + num) > bar->len)) {
-		DEV_VDBG(dev, "bar_idx = %d bar_len = %d offset = %d num = %d val = %d",
-			bar_idx, bar->len, offset, num, *val);
+		DEV_VDBG(dev, "bar_idx = %u bar_len = %u offset = %u num = %u val = %u",
+			bar_idx, (uint32_t)bar->len, (uint32_t)offset, num, *val);
 		return -EFAULT;  /* Bad address */
 	}
